+++ evmv2/contracts/token/VaultTokenV2.sol
@@ -32,7 +32,11 @@ contract VaultTokenV2 is IVaultTokenV2, AccessControlEnumerable, ERC20Burnable {
      *
      * See {ERC20-constructor}.
      */
-    constructor(address _underlying, string memory _name, string memory _symbol) ERC20(_name, _symbol) {
+    constructor(
+        address _underlying,
+        string memory _name,
+        string memory _symbol
+    ) ERC20(_name, _symbol) {
         require(_underlying != address(0), "underlying address is zero");
         _setupRole(DEFAULT_ADMIN_ROLE, _msgSender());
 
@@ -64,12 +68,12 @@ contract VaultTokenV2 is IVaultTokenV2, AccessControlEnumerable, ERC20Burnable {
         if (totalSupply() == 0) {
             return _amount;
         }
-        uint allVToken = totalSupply();
+        uint256 allVToken = totalSupply();
         return _amount.mul(allVToken).div(totalVault);
     }
 
     function getTokenAmount(uint256 _amount) public view override returns (uint256) {
-        uint allVToken = totalSupply();
+        uint256 allVToken = totalSupply();
         if (allVToken == 0) {
             return _amount;
         }
@@ -80,7 +84,11 @@ contract VaultTokenV2 is IVaultTokenV2, AccessControlEnumerable, ERC20Burnable {
         return underlying;
     }
 
-    function deposit(uint256 _fromChain, uint256 _amount, address _to) external override onlyManager {
+    function deposit(
+        uint256 _fromChain,
+        uint256 _amount,
+        address _to
+    ) external override onlyManager {
         uint256 amount = getVaultTokenAmount(_amount);
         _mint(_to, amount);
 
@@ -90,7 +98,11 @@ contract VaultTokenV2 is IVaultTokenV2, AccessControlEnumerable, ERC20Burnable {
         emit DepositVault(underlying, _to, _amount, amount);
     }
 
-    function withdraw(uint256 _toChain, uint256 _vaultAmount, address _to) external override onlyManager {
+    function withdraw(
+        uint256 _toChain,
+        uint256 _vaultAmount,
+        address _to
+    ) external override onlyManager {
         uint256 amount = getTokenAmount(_vaultAmount);
         _burn(_to, _vaultAmount);
